+++ src/component/pages/TodoRegister.jsx
@@ -1,9 +1,7 @@
 // "@emotion/react"には以下が必須
 /** @jsxImportSource @emotion/react */
 
-// import { useState } from "react"
 import { useEffect } from "react"
-// import toast, { Toaster } from "react-hot-toast"
 import { Toaster } from "react-hot-toast"
 import { css } from "@emotion/react"
 import { DayPicker } from "react-day-picker" // react-day-picker：v8.0.１
@@ -36,27 +34,9 @@ export const TodoRegister = () => {
   const inputStyle = css`
     width: 600px;
   `
-  // →→→ ローカルStateは一旦コメントアウト
-  //  初期値Todoタスクのvalueを空にセット、状態を格納する変数setNewTodoをセット
-  // const [newTodo, setNewTodo] = useState("")
-
-  //  初期値Todoの開始日のvalueを空にセット、状態を格納する変数setStartDateをセット
-  // const [startDate, setStartDate] = useState(undefined)
-
-  //  初期値Todoの終了日のvalueを空にセット、状態を格納する変数setEndDateをセット
-  // const [endDate, setEndDate] = useState(undefined)
 
   // グローバルStateの変数群を取り出す。
-  const {
-    newTodo,
-    // setNewTodo,
-    // incompleteTodos,
-    // setIncompleteTodos,
-    startDate,
-    // setStartDate,
-    endDate,
-    // setEndDate,
-  } = useContext(TodoListContext)
+  const { newTodo, startDate, endDate } = useContext(TodoListContext)
 
   // カスタムHookから変数apiPokemonBack,,関数imageFetchを取得
   const { apiPokemonBack, imageFetch } = useImageGet()
@@ -68,42 +48,6 @@ export const TodoRegister = () => {
     imageFetch()
   }, [])
 
-  // todoタスクのテキストボックスで入力した値を保存する
-  // const changeValue = (e) => setNewTodo(e.target.value)
-
-  // 開始日の状態を保存
-  // onDayClickのイベントハンドラーはdayという引数で日程を取得可能
-  // 取得した日程をstartDateの状態を保管
-  // const handleStartDay = (day) => {
-  //   setStartDate(day.toLocaleDateString())
-  // }
-  // 終了日の状態を保存
-  // onDayClickのイベントハンドラーはdayという引数で日程を取得可能
-  // 取得した日程をendDateの状態を保管
-  // const handleEndDay = (day) => {
-  //   setEndDate(day.toLocaleDateString())
-  // }
-
-  // onAddTodoを実施して、incompleteTodosでtodoリスト、完了フラグ、開始日、終了日を格納
-  // const onAddTodo = () => {
-  //   if (newTodo === "") return
-  //   const newTodos = [
-  //     ...incompleteTodos,
-  //     {
-  //       id: incompleteTodos.length,
-  //       todo: newTodo,
-  //       completeFlag: false,
-  //       from: startDate,
-  //       end: endDate,
-  //     },
-  //   ]
-  //   setIncompleteTodos(newTodos) // setIncompleteTodosにnewTodosの状態を登録
-  //   setNewTodo("") // setNewTodoに空の状態を登録
-  //   toast.success("Todoを登録しました.")
-  //   setStartDate(undefined) // 開始日をリセット
-  //   setEndDate(undefined) // 終了日をリセット
-  // }
-
   return (
     <div css={registerStyle}>
       <h2>Todo登録</h2>
@@ -112,7 +56,7 @@ export const TodoRegister = () => {
       <div css={matrixStyle}>
         <div css={registerStyle}>
           <p>１．Todo開始日</p>
-          {/* <DayPicker onDayClick={handleStartDay} /> */}
+          {/* カスタムHookから取得した関数startDayFetchを利用 */}
           <DayPicker onDayClick={startDayFetch} />
           {startDate ? (
             <p> 【Todo開始日】{startDate}</p>
@@ -122,7 +66,7 @@ export const TodoRegister = () => {
         </div>
         <div css={registerStyle}>
           <p>２．Todo完了日</p>
-          {/* <DayPicker onDayClick={handleEndDay} /> */}
+          {/* カスタムHookから取得した関数endDayFetchを利用 */}
           <DayPicker onDayClick={endDayFetch} />
           {endDate ? (
             <p>【Todo終了日】{endDate}</p>
@@ -136,11 +80,10 @@ export const TodoRegister = () => {
         css={inputStyle}
         type="text"
         value={newTodo}
-        // onChange={changeValue}
+        // カスタムHookから取得した関数valueFetchを利用
         onChange={valueFetch}
       />
-      {/* Buttonコンポーネントにアロー関数で関数onAddTodoをPropsで渡す。 */}
-      {/* <Button onClickEvent={() => onAddTodo()}>登録</Button> */}
+      {/* Buttonコンポーネントにアロー関数でカスタムHookから取得した関数todoFetchを渡す。 */}
       <Button onClickEvent={() => todoFetch()}>登録</Button>
       <Toaster />
       <LinkText destination={"/todolist"}>Todo一覧へ</LinkText>
