+++ src/hook/useAddTodo.js
@@ -55,5 +55,6 @@ export const useAddTodo = () => {
     setEndDate(day.toLocaleDateString())
   }
 
+  // 必要な処理、関数を返す
   return { todoFetch, valueFetch, startDayFetch, endDayFetch }
 }
